Split SvgArcSegment
